+++ xlators/cluster/afr/src/afr-self-heal-data.c
@@ -524,7 +524,7 @@ afr_mark_newest_file_as_source(xlator_t *this, unsigned char *sources,
     int i = 0;
     afr_private_t *priv = NULL;
     int source = -1;
-    uint32_t max_ctime = 0;
+    uint64_t max_ctime = 0;
 
     priv = this->private;
     /* Find source with latest ctime */
